+++ GitTest/src/main/java/com/scmaster/gittest/dao/ClipMapper.java
@@ -1,0 +1,13 @@
+package com.scmaster.gittest.dao;
+
+import java.util.List;
+
+import com.scmaster.gittest.vo.Clip;
+
+public interface ClipMapper {
+
+	public int insertClip(Clip clip);
+	
+	public List<Clip> readCLip(String USER_ID);
+	
+}
