+++ GitTest/src/main/resources/mappers/ClipMapper.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.scmaster.gittest.dao.ClipMapper">
+ 
+	<!-- 클립 저장 -->
+	<insert id="insertClip" parameterType="Clip">
+		INSERT INTO SM_CLIP_TB
+		(
+			CLIP_SQ,
+			USER_ID,
+			CONTENT_ID,
+			CONTENT_TYPE_ID,			
+			CLIP_YMD
+			
+		)
+		VALUES
+		(
+			SM_CLIP_TB_CLIP_SQ.nextval
+			, #{USER_ID}
+			, #{CONTENT_ID}
+			, #{CONTENT_TYPE_ID}			
+			, sysdate
+			
+		)
+	</insert>
+	
+	<!-- 클립 읽어오기 -->		
+	<select id="readCLip" parameterType="String" resultType="Clip">
+	select	CLIP_SQ,
+			USER_ID,
+			CONTENT_ID,
+			CONTENT_TYPE_ID,
+			CLIP_YMD
+	from 	SM_CLIP_TB
+	where 	USER_ID=#{USER_ID}
+	order by CLIP_SQ
+	</select>
+	
+	
+</mapper>
