+++ GitTest/src/main/webapp/WEB-INF/views/sc_10.jsp
@@ -240,7 +240,7 @@ var key = "fHPwwCqceBLnLCExz65uYIYEAdiAs6xOwv79o6FcLHh7x6iPmxITE9Wk7TqH1q%2F1%2F
 			$.each(data,function(index,item){
 				console.log(data);
 				var contentId=item.content_ID;
-				var contentTypeId=item.contenttype_ID;
+				var contentTypeId=item.content_TYPE_ID;
 			
 				var url = "http://api.visitkorea.or.kr/openapi/service/rest/KorService/detailCommon?ServiceKey="+key;
 				url += "&mapX&mapY";
@@ -259,7 +259,7 @@ var key = "fHPwwCqceBLnLCExz65uYIYEAdiAs6xOwv79o6FcLHh7x6iPmxITE9Wk7TqH1q%2F1%2F
 			    
 			    
 			    html += '<li>';
-				html += '<a href="SC_07place?CONTENT_ID='+contentId+'&conType='+contentTypeId+'">';
+				html += '<a href="SC_07place?CONTENT_ID='+contentId+'&CONTENT_TYPE_ID='+contentTypeId+'">';
 				
 				html += '<img src='+data.response.body.items.item.firstimage+' width=170 height=190>';
 				html += '<div class="overlay"><h3>'+data.response.body.items.item.title+'</h3></div>';
